Add ComponentLibrary and Component Interfaces (#511)
* Add ComponentLibrary and Component Interfaces

* Component Interface removal

* ComponentLibrary Fat Trimming

* ComponentLibrary Fixes

* ComponentLibrary Fixes

* Add versioning in ComponentLibrary

* ComponentLibrary fixes

* Tidying up ComponentLibrary

* removed version arrays because they were unused;
changed componentVersion check to libraryVersion >= componentVersion;
and some refactors

* we also tell how many component managers were serialized

* Make Serializing Dynamic

I've added a function in wiArchive for retrieving the DATA array
directly. Used to serialize dynamically the archive by choosing to read
datablock or not. Integer hash from string will be used to determine
the existence of the component on the engine's scene or not.

* Make Serialization Dynamic

* archive can jump over unregistered component data; component versioning; refactors;

* component version will be jumped over too

Co-authored-by: Turánszki János <turanszkij@users.noreply.github.com>