Fixing broken hashmaps on linux
Fixes #414

The issue is caused by Linux (and possibly other platforms) implementing the hash function as an identify function for data smaller than 64bit (including `int`).

Fixing only the TerrainGenerator is possible, but the issue could be generated again elsewhere and it would be quite difficult to identify and debug.

Therefore the proposed solution is to replace the `wi::unordered_map` on Linux with another implementation which is still quite fast and does account for the issue of the identify hash function.
The library is https://github.com/martinus/robin-hood-hashing (MIT License)

robin_hood.h:768
```
// return mixed of that, to be safe against identity hash
```