add SHADER_STORAGE_IMAGE_MULTISAMPLE capability and adjust envprobe MSAA
For SPIR-V, dxcompiler translates Texture2DMSArray into ImageMSArray.
The latter is an optional feature for Vulkan, so we have to check it,
and disable features that uses it. Currently, this is only the MSAA mode
for environment probes.

For DX12, Texture2DMSArray is always available, so we enable that
feature unconditionally there.

Further info:
https://registry.khronos.org/vulkan/specs/1.3-khr-extensions/html/vkspec.html#features-shaderStorageImageMultisample