Fix for NaNs in normal for scene capsule collision.
For wi::scene::Scene::Intersects(Capsule), if the overlap depth
is >= the capsule radius, and the line segment that makes up the
spine of the capsule intersects a triangle, then the reported
contact normal would be (NaN, NaN, Nan).  In this special case, the closest
point on the capsule line segment to the triangle ("Center" in the code)
would always be the same as the "bestPoint", which leads up to a divide
by zero in the normal vector calculation.

This fix detects this case, and:
  1) reports the triangle normal when this special case prevents
     calculation of the normal.
  2) corrects the depth calculation for this special case to take into
     the distance from the "Center" point on the spine to the actual
     penetrating endpoint of the spine, projected onto the triangle
     normal.