cmake: replace explicit list of files with globs (#887)
CMake recommends against doing this because some build systems
might not support CONFIGURE_DEPENDS. But currently, the ones
we care about (Make, Ninja) do, and if there's ever a problem,
we can just start listing them again.

Using globs has the advantage that the linux build will not break
when a file is added in VS and somebody forgets to add the file to
cmakelists as well. I've also found some instances were some *.h
files were not listed, which means they weren't copied when installing.