github actions: cancel superseded PR build actions (#1030)
When a PR is updated while the build action hasn't finished yet,
the old action is not automatically cancelled. That build is of
very limited usefulness since its results will not be shown and
are not relevant anymore, so it just wastes resources.

This commit fixes that, now an older build will get cancelled.