linux: don't use getText() in crash handler (#1104)
don't use getText() in crash handler

getText() allocates memory which is a big no no in a signal handler.
Instead, use an internal function that calls us for each entry so we
can write it to the crash log.

This is ugly, but I can't think of a better solution atm.