cmake: add option to use symlinking instead of copying (#1185)
Windows only supports symlinking as admin or in developer mode.
This commit adds a cmake flag to use symlinking instead of copying.
When enabled, it will check for support and output a hint to enable
developer mode when it fails.

By default, it is disabled on Windows, and enabled otherwise.