improve Ccache hit rate
Change the Ccache config so that it has a much higher hit rate
during compilation, see https://ccache.dev/manual/latest.html#_precompiled_headers
for details.

For Clang, always setting -fno-pch-timestamp seems to be ok, so no need
to special case this depending on whether or not Ccache is used.