cmake: handle (non)existance of wiShaderDump.h automatically
We set SHADERDUMP_ENABLED on wiRenderer depending on whether or not
wiShaderDump.h exists. This will cause a recompile if necessary, even
though normally the build system cannot see the has_includes.

Since the compiler flags are different, this will even work with CCache
in direct mode.

This commit also configures CCache to have better hit rate for
precompiled headers as well as re-enabling the (much faster) direct
mode.