sdl: use keyboard layout for letters (#1248)
The windows version already uses the local layout, so that eg. on QWERTZ
keyboard, undo is still CTRL-Z even though Z is not the key next to X.

SDL didn't do this, and forced people to press CTRL-Y instead.

This change also means that AZERTY people will have a hard time navigating
now, but it's consistent with the Windows version.

Fixes #1246